requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
black>=24.1.1
aiogram >=3.18.0
psycopg[binary]
